refactor(engine): rename `action_type` to `action-type` for consistency
