+++ packages/engine/src/domain/action/interfaces.ts
@@ -1,4 +1,4 @@
-import type { ActionType } from "@/domain/action/action_type";
+import type { ActionType } from "@/domain/action/action-type";
 import type { ICoordinate } from "@/math/coordinate";
 
 export interface IAction {
